requirements.txt
alembic==1.5.7
appdirs==1.4.4
astroid==2.5
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
protobuf==3.15.6
pylint==2.7.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.0
toml==0.10.2
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
